[go] Create a components directory and move all existing components under it
