+++ go/cmd/controllermgr/BUILD.bazel
@@ -11,9 +11,8 @@ go_library(
         "//go/base/config:go_default_library",
         "//go/base/env:go_default_library",
         "//go/base/zapfx:go_default_library",
+        "//go/components/ray:go_default_library",
         "//go/controllermgr:go_default_library",
-        "//go/controllers/raycluster:go_default_library",
-        "//go/controllers/rayjob:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_go_logr_zapr//:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
