+++ go/components/ray/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["module.go"],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/components/ray",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//go/components/ray/cluster:go_default_library",
+        "//go/components/ray/job:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
+    ],
+)
