+++ go/components/ray/cluster/module.go
@@ -1,23 +1,23 @@
-package rayjob
+package cluster
 
 import (
 	"go.uber.org/fx"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 
 	"github.com/michelangelo-ai/michelangelo/go/base/env"
-	"github.com/michelangelo-ai/michelangelo/go/controllers/raycluster"
 	rayv1 "github.com/ray-project/kuberay/ray-operator/pkg/client/clientset/versioned/typed/ray/v1"
 )
 
 var (
 	// Module FX
 	Module = fx.Options(
+		fx.Provide(newConfig),
 		fx.Invoke(register),
 	)
 )
 
 func register(
-	conf raycluster.Config,
+	conf Config,
 	env env.Context,
 	mgr manager.Manager,
 ) error {
