+++ go/components/ray/job/BUILD.bazel
@@ -3,27 +3,25 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "config.go",
         "controller.go",
         "module.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/controllers/raycluster",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/components/ray/job",
     visibility = ["//visibility:public"],
     deps = [
         "//go/api/utils:go_default_library",
         "//go/base/env:go_default_library",
+        "//go/components/ray/cluster:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_go_logr_logr//:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//apis/ray/v1:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//pkg/client/clientset/versioned/typed/ray/v1:go_default_library",
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_apimachinery//pkg/types:go_default_library",
         "@io_k8s_sigs_controller_runtime//:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/log:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/manager:go_default_library",
-        "@org_uber_go_config//:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
 )
@@ -33,15 +31,14 @@ go_test(
     srcs = ["controller_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//go/controllers/utils/testutils:go_default_library",
+        "//go/components/testfakes:go_default_library",
+        "//proto/api:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//apis/ray/v1:go_default_library",
-        "@com_github_ray_project_kuberay_ray_operator//pkg/client/clientset/versioned/scheme:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//pkg/client/clientset/versioned/typed/ray/v1/fake:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/api/resource:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
         "@io_k8s_apimachinery//pkg/types:go_default_library",
