+++ go/components/testfakes/fake_client_wrapper.go
@@ -1,4 +1,4 @@
-package testutils
+package testfakes
 
 import (
 	"context"
