+++ go/components/testfakes/reactor_manager.go
@@ -1,4 +1,4 @@
-package testutils
+package testfakes
 
 import (
 	"sync"
